+++ src/Overlook/FixedSimBroker.h
@@ -48,13 +48,13 @@ struct FixedSimBroker {
 	int signal[SYM_COUNT];
 	int proxy_id[SYM_COUNT];
 	int proxy_base_mul[SYM_COUNT];
-	long double balance = 0.0;
-	long double equity = 0.0;
-	long double part_balance = 0.0;
-	long double part_equity = 0.0;
-	long double begin_equity = 0.0;
-	long double profit_sum = 0.0;
-	long double loss_sum = 0.0;
+	double balance = 0.0;
+	double equity = 0.0;
+	double part_balance = 0.0;
+	double part_equity = 0.0;
+	double begin_equity = 0.0;
+	double profit_sum = 0.0;
+	double loss_sum = 0.0;
 	double spread_points[SYM_COUNT];
 	double free_margin_level = 0.80;
 	double leverage = 1000.0;
@@ -67,7 +67,7 @@ struct FixedSimBroker {
 	void Reset();
 	double RealtimeBid(const Snapshot& snap, int sym_id) const;
 	double RealtimeAsk(const Snapshot& snap, int sym_id) const;
-	long double GetCloseProfit(int sym_id, const FixedOrder& o, const Snapshot& snap) const;
+	double GetCloseProfit(int sym_id, const FixedOrder& o, const Snapshot& snap) const;
 	void OrderSend(int sym_id, int type, double volume, double price, const Snapshot& snap);
 	void OrderClose(int sym_id, double lots, FixedOrder& order, const Snapshot& snap);
 	void CloseAll(const Snapshot& snap);
@@ -75,8 +75,8 @@ struct FixedSimBroker {
 	bool Cycle(const Snapshot& snap);
 	void RefreshOrders(const Snapshot& snap);
 	
-	long double AccountEquity() const {return equity;}
-	long double PartialEquity() const {return part_equity;}
+	double AccountEquity() const {return equity;}
+	double PartialEquity() const {return part_equity;}
 	double GetDrawdown() const {double sum = profit_sum + loss_sum; return sum > 0.0 ? loss_sum / sum : 1.0;}
 	void SetSignal(int sym_id, int sig) {ASSERT(sym_id >= 0 && sym_id < SYM_COUNT); signal[sym_id] = sig;}
 	int GetSignal(int sym_id) const {ASSERT(sym_id >= 0 && sym_id < SYM_COUNT); return signal[sym_id];}
